refactor: rename logging -> io + add more funcs
