+++ src/caelestia/subcommands/resizer.py
@@ -7,7 +7,7 @@ from pathlib import Path
 from typing import Any, Dict, Optional
 
 from caelestia.utils import hypr
-from caelestia.utils.logging import log_message
+from caelestia.utils.io import log_message
 from caelestia.utils.paths import user_config_path
